improvement(Add ability to run library both on server and browser): We removed Axios and are now usi
There were problems running this library on server and browser environments. Sometimes axios was
using the wrong adapter. We removed axios and are now using fetch for our requests. We recommend to
polyfill fetch in node environments through the cross-fetch library that is available via npm.

BREAKING CHANGE: Since fetch is used for requesting data from the interfaces it is necessary to
polyfill fetch, when you are using this library in node environments.