feat(navigation): we added support for multi-language projects
you can now pass the initial path of your application to fetchNavigation. it will be used to
determine the correct language and will return the full nav-tree. if no initialPath or "/" is passed
the defaultLocale will be used.

BREAKING CHANGE: the interface of the fetchNavigation changed. You now have to pass the initial path
of your application and a fallback defaultLocale to the method.