feat(custom-mapper): we've added a customMapper option that will override the default mapping
You can now pass a customMapper-function. This will override all existing mapping functionality if
you return anything from it. If you do not return something, the default mapping will take
place.
This customMapper will be used, when single Data-Entries are mapped. A useful util-object is
passed as well, which enables you to register items that should be loaded in one single request to
the CaaS (Images for example) or even fetch other documents from the the CaaS. Since the
customMapper must return a Promise async operations are no problem at all.