feat(errors): we are now throwing 404 and 401 errors / renamed fetchPage to fetchElement
We are now throwing 404 / 401 errors, so that you can better react to specific errors and are
notified about wrong api-keys for example. We also renamed the fetchPage to fetchElement to better
mirror its functionality. This method can now return a GCAPage, a Dataset, a Page and an Image.
fetchElement and fetchByFilter will now accept additionalParams that will be passed to the
restheart-api. You can pass all available restheart-params. Please be aware that both methods will
skip the data-mapping, when the keys parameter is added, since we cannot ensure, that all of the
required properties are available.

BREAKING CHANGE: - renamed fetchPage to fetchElement. This method will now return mapped GCAPage /
Image / Dataset as well
-Methods will now throw errors, when an element is not found or the request
was not authorized