feat: enabled usage of remote media (#30)
This feature enables the use of remote media.
However, it is only possible to use this feature if at least version 2021-06 is installed on the FirstSpirit server.

A new variable must be specified in the environment variables so that the remote project is registered.
This is JSON which contains the name of the project as the key and an object as the value. This object must have an attribute where the key is `id` and the value the project id and a language, for this the key is `locale` and the value is the language abbreviation e.g. `en_GB`. Multiple projects can also be specified here.

Attention!
Currently the remote media can only be fetched with the language provided in the object.

If an element is requested via fetchElement or fetchByFilter and an element contains a remote medium, this remote medium is resolved and all available resolutions are available.
If you want to access a remote medium directly, the methods fetchElement and fetchByFilter offer the possibility to enter the project name to access the medium.

This all is documented in the Readme.

Also it has been made easier for developers to test the FSXA API. Since now in the folder dev a method is added, which makes it possible to write the response into a file and these appear therefore not only in the console. But this is all optional.