feat: add FSXAProxy and FSXARemoteApi, remove FSXAApi, enable navigationFilter (#44)
We've removed the FSXAApi Class and added two new classes. The FSXAProxyApi for the proxy mode and
FSXARemoteApi for the remote mode. The logic of the main methods remains the same, but they are
called differently. In addition you can configure a navigationFilter and preFilterFetch method to
hide navigationdata.

BREAKING CHANGE: The FSXAApi class was removed and new classes FSXAProxyApi and FSXARemoteApi was
added. Please read the migration guide in the README for more details.