feat: add FSXAProxy, FSXARemoteApi and navigationFilter utilities (#58)
* feat: add FSXAProxy and FSXARemoteApi, remove FSXAApi, enable navigationFilter (#44, #55, #57)
* refactor: add TypeDoc and minor code improvements (#47, #49)

We have removed the FSXAApi Class and added two new classes. The FSXAProxyApi for the proxy mode and
FSXARemoteApi for the remote mode. The logic of the main methods remains the same, but they are
called differently. In addition you can configure a navigationFilter and preFilterFetch method to
hide navigation data you do not want to transfer. We have added TypeDoc so developers have a well documented API and have improved logging.

BREAKING CHANGE: The FSXAApi class was removed and new classes FSXAProxyApi and FSXARemoteApi was
added. Please read the [migration guide in the CHANGELOG](https://github.com/e-Spirit/fsxa-api/blob/alpha/CHANGELOG.md#migration-guide) for more details.